Split sensors to front/back of car
Sensors on the same CAN ID cannot be sent from oppisite sides of the
car, need to be grouped into which MC is sending their data (front/back)

Encapsulate CSV reading into class

This will allow for us to pre-load the sensor dict to speed up parsing
individual data frames.

Re-arrange dict format

MSB should be places before LSB as this is how it is formatted in the
CAN messages.

Add sample usage code and doccumentation

Sample code can be used as a quick unit test and to ensure that the CSV
format is valid