Feature add data logging (#7)
* Add functionality for reading serial data

Serial data is read from the CANdapter and sent to all clients connected
on websockets.
Improved client view shows data in realtime graph.

* Fixed linting issues

Single line was too long...

* Import full .gitignore file

Fixes #1

* Add README.md

Our repo doesn't have a README, this commit adds one.

* Add inital data logging

Output data logs to the ./logs directory and mimic the styel of past
logs made with the wifican.js file to allow for legacy Matlab
processing.

* Add final data formatting and IP address

Add final ip configuration for use with router on car

* Remove clientside code

This code is unessicary, replaced by KX dashboard

* Add router configuration files

Add backup of router configuration. Can be loaded by the router control
panel.

* Fix linting issues.

* Added a period

* Fix debug messages and double comment

Remove debugging print and double comment to make ryan happy so he can
merge this old pull request. This is important because this code is 3
months old and needs to be put on master eventually.

* Add initial sensor dictionaly with CAN IDs

This is a somewhat complete list of snesors on the DCAN bus that will be
sent over the wireless telemetry server. We need to add additional
sensors on the DCAN such as BMS info and finalize the IDs for each
sensor.

* Add potential IDs and byte indicies.

First revision of CAN information for new sensors

* Update message ID formatting

* Remove client HTML code

Static HTML client is no longer needed for viewing data

* Update tire temp byte length

Ture temp is being sent as a double value, needs 8 bits to represent
it.

* Adds a function to read sensor data from dictionary

data_to_dict Takes in data and reformats it to make it more readable. As long as the SensorDict.csv stays in the same format, this probably won't need many changes in the future.

* Add print formatting for excel/google sheets

Adds an option for a console printout of the CSV file to be pasted into
the excel sheet that keeps track of all the CAN IDs

* Adds function to save sensor data from a list

Also updated data_to_dict to include "timestamp" in its output. This is so that these functions can be used together, both using timestamps.

* Output correct start/stop bits for can IDs

* Update SensorDict.py to coding standards

Used autopep8 to update the file.

* Updates SensorDict.py with line limit 80

Limit was not 80 when I ran autopep before.

* Change CAN message format for IMU

* Fix linting issues

Travis plz

* Add support for parsing multiple data points

Will import data from 'SensorData.csv' and parse incoming data packets
based on the information peovided in the file. This file is used to
define each sensor's address and start/end bytes used across the CAN bus
and wireless Telemetry server.

* Split sensors to front/back of car

Sensors on the same CAN ID cannot be sent from oppisite sides of the
car, need to be grouped into which MC is sending their data (front/back)

Encapsulate CSV reading into class

This will allow for us to pre-load the sensor dict to speed up parsing
individual data frames.

Re-arrange dict format

MSB should be places before LSB as this is how it is formatted in the
CAN messages.

Add sample usage code and doccumentation

Sample code can be used as a quick unit test and to ensure that the CSV
format is valid

* Fix linting issues

* Fix typos mentioned in #7

* Add C-compatiable names

We need additional names for generating the C header files.

* Script to upate SensorDict from Google Sheets

Requires pandas and requests.
Saves the sensor info in the same format as SensorDict.py in file `output_file`
Ignores the first `ignore_rows` rows

* update-SensorDict script adds comments

The comments column in the
spreadsheet will be used. The first byte
of each sensor will have a comment if
applicable.

* Add file to generate C headers

This script will read a SensorDict file
and generate C style headers for the
names and bytes.

* Format CANID names to work with header file gneration tool

* Fix formatting errors and duplicate message ids.

* Fix linting errors...