Add canid file structure (#15)
* Move CAN related files to new directory

This will help seperate the two parts of the project.

* Add documentation to generated header file

This will make time-stamping in main module easier to manage if each
header file is dated at the top.

* Ignore nan rows from google sheets

This will allow for the google sheets parsing to work when some columns
are not formatted correctly/unable to parse currently.

* Makes update script use iterrows

I did not have it iterating through the
dataframe correctly. This will be better.

* Makes header script use iterrows

Now using a better way to iterate
through the file. Also added formatting
to make the strings more readable.

* Combines header gen. and sensordict updater

One script will now update
SensorDict.csv and CANID.h. It will
ignore rows that don't have correctly
formatted names.

* Change CSVs to use bits instead of bytes

Also added flags from the spreadsheet,
which are found when there is a newline
in the cell.

* Delete old CANID.h