Add pipenv to manage dependencies, solves #9 (#10)
* Add pipenv to manage dependencies, solves #9

pipenv ensures that you can run the python files with the correct pip
modules on different systems. While not as complete as a docker
solution, this will solve our needs for this simple project.

* Breakdown of how to use pipenv in readme

A simple description of how to use pipenv for this project. I assume python has been added to PATH for this.

* Update to Python 3.7.4
Added "Pandas" library to Pipfile.

* Add Requests and mysql-connector libraries
Requests for accessing Google Drive spreadsheet and mysql for database.